refractor: Refractor closely related code as classes
