+++ src/data/tasks.txt
@@ -1,8 +1,1 @@
-T | 1 | sleep
 T | 0 | read
-T | 0 | quiz
-D | 0 | study  | 2024-12-12T13:00
-E | 0 | talk  | 2024-01-12T13:00 | 2024-02-23T15:00
-D | 0 | study  | 2024-03-12T11:00
-D | 0 | study  | 2024-12-12T10:00
-T | 0 | sleep
