+++ src/main/java/ByeCommand.java
@@ -1,0 +1,7 @@
+public class ByeCommand extends Command {
+
+    @Override
+    public void executeCommand(TaskList taskList, Ui ui, Storage storage) {
+        ui.bye();
+    }
+}
