+++ src/main/java/ListCommand.java
@@ -1,0 +1,6 @@
+public class ListCommand extends Command {
+    @Override
+    public void executeCommand(TaskList taskList, Ui ui, Storage storage) {
+        taskList.printTask();
+    }
+}
