+++ src/main/java/NoTimeline.java
@@ -9,7 +9,7 @@ public class NoTimeline extends ToothlessExceptions{
      * @param instruction the format for the task type
      */
     public NoTimeline(String taskType, String instruction) {
-        super("The timing for " + taskType +  "is not specified! :o\n" +
+        super("The timing for " + taskType +  " is not specified! :o\n" +
                 "Please enter a valid timeline.\n\n" +
                 "The format for " + taskType + " is: " + instruction + "\n" +
                 "*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*\n");
