+++ src/main/java/UnknownCommand.java
@@ -1,0 +1,7 @@
+public class UnknownCommand extends Command {
+
+    @Override
+    public void executeCommand(TaskList taskList, Ui ui, Storage storage) {
+        ui.unknownCommand();
+    }
+}
